requirements.txt
asgiref==3.8.1
boto3==1.37.6
botocore==1.37.6
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==4.2.18
django-allauth==65.3.1
django-countries==7.2.1
django-crispy-forms==2.3
django-storages==1.14.5
gunicorn==23.0.0
jmespath==1.0.1
oauthlib==3.2.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
requests-oauthlib==2.0.0
s3transfer==0.11.4
sqlparse==0.5.3
stripe==11.5.0
whitenoise==6.9.0
